pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "vegetablesellerbot"
version = "0.1.0"
[tool.poetry.dependencies]
google-nucleus = "^0.5.0"
gspread = "^3.4"
python = "^3.7"
cryptg = "^0.2.0"
telethon = "^1.11"
pymongo = "^3.10"
dnspython = "^1.16"
flask = "^1.1"
[tool.poetry.dev-dependencies]
